2d example + modified beam definition
Made a 2D LWFA example to speed up testing.

In the process it became necessary to add an attribute to the Gaussian Beam.  In 2D Cartesian (or r-z)

Here I assume the distribution is Gaussian and the density is related to the physical particles via the relation

N = density * product(sigma) * (2*pi)^(3/2)

In the 3D example, the total beam particles is 10^8 and the RMS width is 1 micron in all directions, this corresponds to a density of 6.35*10^24 m^(-3).