mv attribute docstrings
The format before this commit has been proposed in PEP 224, which has
been rejected.
Aim of this format is to avoid duplicating the list of attributes in the
class body and class docstring.

However, the old format is only understood by some tools, so we return
to plain PEP 257 docstrings.

Note: An alternative would be to add a __doc_ATTRNAME__ string, which
would only exist as convention. (Could be enforced inside PICMI though.)